feat(Index): Switched stateless functional component for full component
Index page needs to keep track of a very basic state, namely which item has been chosen in the
Select menu. By upating the state with that choice, a re-render of the Component is triggered. The
updated state is used to filter the Product data displayed by checking if the current state value is
present in each of the Product's sizes property. If it is, the Product is dispalyed, otherwise it
isn't. If the user selects 'all', all Products are displayed.