feat(Articles): Added basic transition tied to state
Instead of showing all sizes (default state) briefly before the localStorage value is loaded in, the
opacity of Articles Component is set to zero. When the componentDidMount method is fired a state
change adds a class to the wrapping div which is tied to a CSS transition to set opacity to 1. This
just makes the experience a little nicer.